+++ cmd/flag.go
@@ -9,15 +9,17 @@ import (
 )
 
 type flag struct {
-	fields []string
-	greps  []string
-	output string
+	fields  []string
+	group   string
+	output  string
+	selects []string
 }
 
 func (f *flag) Init(cmd *cobra.Command) {
 	cmd.PersistentFlags().StringSliceVarP(&f.fields, "field", "f", nil, "Fields the output lines should contain only.")
-	cmd.PersistentFlags().StringSliceVarP(&f.greps, "grep", "g", nil, "Grep for lines based on the given key:val regular expression.")
+	cmd.PersistentFlags().StringVarP(&f.group, "group", "g", "", "Group logs by inserting an empty line after the group end.")
 	cmd.PersistentFlags().StringVarP(&f.output, "output", "o", "json", "Output format, either json or text.")
+	cmd.PersistentFlags().StringSliceVarP(&f.selects, "select", "s", nil, "Select lines based on the given key:val regular expression.")
 }
 
 func (f *flag) Validate() error {
@@ -32,22 +34,6 @@ func (f *flag) Validate() error {
 		}
 	}
 
-	// Validate -g/--grep flags.
-	for _, g := range f.greps {
-		split := strings.Split(g, ":")
-
-		if len(split) != 2 {
-			return microerror.Maskf(invalidFlagsError, "%#q must have format key:val", g)
-		}
-
-		for _, s := range split {
-			_, err := regexp.Compile(s)
-			if err != nil {
-				return microerror.Mask(err)
-			}
-		}
-	}
-
 	// Validate -o/--output flag.
 	{
 		if f.output == "" {
@@ -55,6 +41,22 @@ func (f *flag) Validate() error {
 		}
 		if f.output != "json" && f.output != "text" {
 			return microerror.Maskf(invalidFlagsError, "-o/--output must either be text or json")
+		}
+	}
+
+	// Validate -s/--select flags.
+	for _, s := range f.selects {
+		split := strings.Split(s, ":")
+
+		if len(split) != 2 {
+			return microerror.Maskf(invalidFlagsError, "-s/--select must have format key:val")
+		}
+
+		for _, s := range split {
+			_, err := regexp.Compile(s)
+			if err != nil {
+				return microerror.Mask(err)
+			}
 		}
 	}