+++ CHANGELOG.md
@@ -14,6 +14,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Make multiple fuzzy grepping generic.
 - Implement colourized field formatting.
 - Add usage info to readme.
+- Rename `-g` and `--grep` flags to `-s` and `--select`.
+- Group output logs using regular expressions via the `-g` or `--group` flag.